requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
googlesearch-python
python-dotenv
httpx
duckduckgo-search
